requirements.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.0.0
qrcode[pil]>=7.4.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1
